Add pefInterpolation.py recipe to the experiments #43
Changes pefInterpolation function calling from fullInterpolationAndStack
experiments.

The pefInterpolation.py recipe is now part of official Madagascar
package. Because of that, SConstruct can call it from local Madagascar
installation.
